server/requirements.txt
fastapi
uvicorn
pydantic-settings
requests
redis
asyncpg
sqlalchemy
httpx
psycopg2-binary
